release: specify target repository
This commit resolved the following error.

```
failed to run git: fatal: not a git repository (or any of the parent directories): .git
```

The cause of the above error gh command doesn't found target repository because of .git does not exist.
So, we specify the target repository by `--repo` option in this commit.