release: specify target repository (#26)
This commit resolved the following error.

https://github.com/groonga/docker/actions/runs/15405611703/job/43347754316#step:4:10

```
failed to run git: fatal: not a git repository (or any of the parent directories): .git
```

The cause of the above error `gh` command doesn't found target
repository because of `.git` does not exist. So, we specify the target
repository by `--repo` option in this commit.

---------

Co-authored-by: Abe Tomoaki <abe@enzou.tokyo>